build(meson): generate CMake package config files
The cmake/config.cmake.in file uses the 'install(EXPORT)` CMake feature,
that is not (yet?) supported by Meson (see
https://github.com/mesonbuild/meson/issues/7632#issuecomment-704932266).

To work around this issue I created a simple CMake config file that
defines a single IMPORTED target.

It is really basic, but since xbyak is a header-only library defining a
single target works without issues.

This is useful because with this feature when building the library using
Meson, CMake users will be able to still use `find_package(xbyak)`, and
are not forced to use the FindPkgConfig module