ci: use containers
GitHub's CI default runners are full of all kinds of software, and this
can cause hard to reproduce errors. Luckly, GitHub Actions supports
running jobs in a container image, so that the build environment is
minimal and reprocucible.

This patch changes the main CI job to run in a Debian Testing container,
fixing the issues described in
<https://github.com/herumi/xbyak/commit/cfb1127c3bf28d55e68b367a0a6944829a5c025d>.

As Debian Testing is a semi-rolling relase distribution, and nasm 2.15
is shipped by default, I've also modified the script to install nasm
from apt instead of building it from source.