add detection of xsave
The XSAVE cpuid flag determines if the processor supports the XSAVE
instructions(`xsave`, `xrestor`, `xsetbv`, `xgetbv`).

The OSXSAVE cpuid flag determines if the _operating system_ has XSAVE
**enabled**. Userland code has to use OSXSAVE for features such as AVX
and AVX512 but it still matters to know if the processor itself supports
it but not the operating system such as some virtual-machine contexts.